Move HF batch streamer to handler_utils
